refactor(api): extract user service database access to repository
