+++ api/app/Services/PatientService.php
@@ -2,53 +2,27 @@
 
 namespace App\Services;
 
-use App\Models\Patient;
+use App\Repositories\PatientRepository;
 use App\Util\PaginationUtil;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Response;
 
-class PatientService
+final readonly class PatientService
 {
+    public function __construct(
+        private PatientRepository $patientRepository
+    )
+    {
+
+    }
+
     public function list(array $parameters): Response
     {
-        $patients = $this->getPatients($parameters);
+        $builder = $this->patientRepository->findPatientsByQueryFields($parameters);
 
-        return new Response($patients, Response::HTTP_OK);
-    }
+        $paginator = $builder->paginate($parameters['per_page'], ['*'], 'page', $parameters['page']);
 
-    private function getPatients(array $parameters): array
-    {
-        $query = Patient::query()->join('users', 'patients.user_id', '=', 'users.id')
-            ->select('patients.uuid', 'users.name', 'users.cpf', 'users.email',
-                'patients.cns', 'patients.created_at');
+        $pageData = PaginationUtil::extractData($paginator);
 
-        $query = $this->filterQueryByFields($query, $parameters);
-        $paginator = $query->paginate($parameters['per_page'], ['*'], 'page', $parameters['page']);
-
-        return PaginationUtil::extractData($paginator);
-    }
-
-    private function filterQueryByFields(Builder $query, array $parameters): Builder
-    {
-        $blackList = ['offset', 'page', 'limit'];
-        $patientBlackList = ['name', 'cpf', 'email'];
-
-        foreach ($parameters as $key => $parameter) {
-            if (strlen($key) === 5 && str_starts_with($key, 'name') && !in_array($parameter, $blackList)) {
-                $number = substr($key, 4, 5);
-
-                if (array_key_exists("value$number", $parameters)) {
-                    $query =
-                        $query->whereHas('user', function ($query) use ($parameter, $parameters, $number, $patientBlackList) {
-                            $searchValue = $parameters["value$number"];
-                            $table = in_array($parameter, $patientBlackList) ? 'users' : 'patients';
-
-                            $query->whereRaw("$table.$parameter LIKE '$searchValue%'");
-                        });
-                }
-            }
-        }
-
-        return $query;
+        return new Response($pageData, Response::HTTP_OK);
     }
 }
